scripts/pyproject.toml
[project]
name = "ear-trainer-scripts"
version = "0.1.0"
description = "Audio generation scripts for ear-trainer"
requires-python = ">=3.10"
dependencies = [
    "praat-parselmouth>=0.4.0",
    "soundfile>=0.12.0",
    "numpy>=1.24.0",
]
